+++ bin/check-ssl-host.rb
@@ -1,5 +1,6 @@
 #!/usr/bin/env ruby
-# encoding: UTF-8
+# frozen_string_literal: false
+
 #  check-ssl-host.rb
 #
 # DESCRIPTION:
@@ -42,7 +43,7 @@ require 'socket'
 # Check SSL Host
 #
 class CheckSSLHost < Sensu::Plugin::Check::CLI
-  STARTTLS_PROTOS = %w(smtp imap).freeze
+  STARTTLS_PROTOS = %w[smtp imap].freeze
 
   check_name 'check_ssl_host'
 
@@ -104,7 +105,7 @@ class CheckSSLHost < Sensu::Plugin::Check::CLI
          long: '--starttls PROTO'
 
   def get_cert_chain(host, port, address, client_cert, client_key)
-    tcp_client = TCPSocket.new(address ? address : host, port)
+    tcp_client = TCPSocket.new(address ? address : host, port) # rubocop:disable Style/RedundantCondition
     handle_starttls(config[:starttls], tcp_client) if config[:starttls]
     ssl_context = OpenSSL::SSL::SSLContext.new
     ssl_context.cert = OpenSSL::X509::Certificate.new File.read(client_cert) if client_cert
@@ -139,6 +140,7 @@ class CheckSSLHost < Sensu::Plugin::Check::CLI
 
     status = socket.readline
     return if /^220 / =~ status
+
     critical "#{config[:host]} - did not receive SMTP 220 in response to STARTTLS"
   end
 
@@ -151,6 +153,7 @@ class CheckSSLHost < Sensu::Plugin::Check::CLI
 
     status = socket.readline
     return if /^a001 OK Begin TLS negotiation now/ =~ status
+
     critical "#{config[:host]} - did not receive OK Begin TLS negotiation now"
   end
 
@@ -158,7 +161,7 @@ class CheckSSLHost < Sensu::Plugin::Check::CLI
     # Expiry check
     days = (cert.not_after.to_date - Date.today).to_i
     message = "#{config[:host]} - #{days} days until expiry"
-    critical "#{config[:host]} - Expired #{days} days ago" if days < 0
+    critical "#{config[:host]} - Expired #{days} days ago" if days < 0 # rubocop:disable Style/NumericPredicate
     critical message if days < config[:critical]
     warning message if days < config[:warning]
     ok message