+++ CHANGELOG.md
@@ -11,6 +11,9 @@ This CHANGELOG follows the format listed [here](https://github.com/sensu-plugins
 - Travis build automation to generate Sensu Asset tarballs that can be used n conjunction with Sensu provided ruby runtime assets and the Bonsai Asset Index
 - Require latest sensu-plugin for [Sensu Go support](https://github.com/sensu-plugins/sensu-plugin#sensu-go-enablement)
 
+### Changed
+- Updated bundler development depedency to '~> 2.1'
+
 ## [2.0.1] - 2018-05-30
 ### Fixed
 - `check-ssl-qualys.rb`: Fixed typo and removed timeout `-t` short option replacing it with `--timeout` as per previous changelog. `-t` conflicts with the short option for `--time-between`