+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This CHANGELOG follows the format listed [here](https://github.com/sensu-plugins
 ### Added
 - Travis build automation to generate Sensu Asset tarballs that can be used n conjunction with Sensu provided ruby runtime assets and the Bonsai Asset Index
 - Require latest sensu-plugin for [Sensu Go support](https://github.com/sensu-plugins/sensu-plugin#sensu-go-enablement)
+- New Check plugin `check-ssl-root-issuer.rb` with alternative logic for trust anchor verification.
 
 ## [2.0.1] - 2018-05-30
 ### Fixed